requirements.txt
opencv-python
flask
mako
ruamel.yaml